+++ petitparser/lib/parser.dart
@@ -43,3 +43,4 @@ export 'src/parser/repeater/possessive.dart';
 export 'src/parser/repeater/repeating.dart';
 export 'src/parser/repeater/separated_by.dart';
 export 'src/parser/repeater/unbounded.dart';
+export 'src/parser/utils/failure_joiner.dart';